refactor: Changed relapse and update commands to do channel checks natively
Changed the relapse and update commands to do channel checks nativley using the discord.py wrapper. Now inside the decrators of the commands, there is an additional decorator that leverages a function in the utils file to check if the server has a channel limit, and if it does make sure messages are being sent in the correct channel.