refactor: Changed location of motivation command to its own cog
Changed the location of the motivation command to its own cog. Previously it was stored in the development  cog, this cog was made for developers to test and debug the bot as of when it is needed. Aswell as this various bug fixes were impliment that addressed probems in previously commited code