feat: Created the !setup_roles command
Created the !setup_roles command. This command does two things, firstly it creates the common roles used in discord servers, names and colours included. Then it inserts these roles into the databse for future referance. This command currently does have litations, the main one being an admin or moderator can run it multiple time and this can break functionaility of the `update_roles` function in the streaks.py cog, aswell as this the roles can be deleted and this could also potentially cause bugs. Im going to create an issue reflectig these, and they need to be fixed before this is merged into the master branch. Aswell as this the doc strings for these commands also need to be inserted