feat: Added overide and safegaurd to !setup_roles command
Added override and safeguard to the !setup_roles command. Now, if there are already roles in the database associated with the server the command is being sent in, the bot will prompt the author "are you sure you want to do this". It then offers them an override option if required.  The override option deleted all the previous roles and then deleted all the entries in the database associated with those roles (this second part could be subject to change)