refactor: Changed logging so SQL queries use logging module rather than SQLalchemy
Made various changes to the logging functionality of the bot. The main change is that the program no longer relies on SQLalchemy to log SQL queries, it now does this natively using the logging functionality added in the last release. All these functions are done during the class function used to call the SQL query, and call out to a separate function to do so (outsource due to standard pattern in text). This change made it so that `echo=True` could be removed from the main.py file, and the default arg for that changed in the database init function. The other main change is that now the logger outputs logs to bot the .log file and the terminal now

Closes issue #22