feat: Added invite link command
Added a command for bot to send invite link. This comes with two caveats, the first one being, this link must be added to the settings.json file as it is instance specific. The second one being, before the bot is made public, the link must be updated so the bot does not have admin privilages