fix: Applied various fixes to on_guild_join listener
Applied various fixes to on_guild_join listener. Firstly it created an info log that the bot has been added to a server. It also attempts to send a message to the bot owner first, failing that the guild channel w/ the name "general". Aswell as this on the databse level it checks if the guild is already in the database

Closes issue #29