fix: Fixed error in get_owned_roles function
fixed an error in the get_owned_roles function. Function unnececarily tried to fetch the role using the discord api despite the fact it already had the role. As the function requested the API to return the role using the role itself as the input, this caused it to return "None". This in turn meant the bot couldnt delete roles from the user of which they have surpassed. This fix surpasses that, but cannot deal with users who still have the wrong roles